core: restore global controllers and update di
