+++ lib/features/auth/controller/auth_controller.dart
@@ -1,53 +1,40 @@
-import 'dart:convert';
-
-import 'package:flutter_boilerplate/features/auth/repository/auth_repo.dart';
-import 'package:flutter_boilerplate/data/model/user/user.dart';
-import 'package:flutter_boilerplate/data/model/user/role.dart';
+import 'package:flutter_boilerplate/features/auth/model/user.dart';
+import 'package:flutter_boilerplate/features/auth/service/auth_service.dart';
+import 'package:flutter_boilerplate/features/auth/model/role.dart';
 import 'package:get/get.dart';
 
-class AuthController extends GetxController implements GetxService {
-  final AuthRepo authRepo;
-  AuthController({required this.authRepo}) {
-    loadUser();
-  }
+class AuthController extends GetxController {
+  final AuthService authService;
+  AuthController({required this.authService});
 
-  UserModel? _currentUser;
-  UserModel? get currentUser => _currentUser;
+  UserModel? get currentUser => authService.currentUser;
 
   void loadUser() {
-    final jsonString = authRepo.getUser();
-    if (jsonString != null) {
-      _currentUser = UserModel.fromJson(jsonDecode(jsonString));
-    }
+    authService.loadUser();
+    update();
   }
 
   Future<void> signUp(UserModel user) async {
-    await authRepo.signUp(jsonEncode(user.toJson()));
-    _currentUser = user;
+    await authService.signUp(user);
     update();
   }
 
   Future<void> login(String email, String password) async {
-    final response = await authRepo.login(email, password);
-    if (response.isOk) {
-      loadUser();
-    }
+    await authService.login(email, password);
     update();
   }
 
-  /// Remove the login flag/token.
   Future<void> logout() async {
-    await authRepo.logout();
+    await authService.logout();
+    update();
   }
 
-  /// Whether the user is currently logged in.
-  bool get isLoggedIn => authRepo.isLoggedIn();
+  bool get isLoggedIn => authService.isLoggedIn;
 
-  bool get isOwner => currentUser?.role == UserRole.owner;
+  bool get isOwner => authService.isOwner;
 
   Future<void> saveUser(UserModel user) async {
-    _currentUser = user;
-    await authRepo.saveUser(jsonEncode(user.toJson()));
+    await authService.saveUser(user);
     update();
   }
 }
