+++ lib/features/auth/model/role.dart
