+++ lib/features/auth/model/user.dart
@@ -1,4 +1,4 @@
-import '../wallet/wallet.dart';
+import 'package:flutter_boilerplate/features/finance/model/wallet.dart';
 import 'role.dart';
 
 class UserModel {
