+++ lib/features/auth/service/auth_service.dart
@@ -1,0 +1,51 @@
+import 'dart:convert';
+
+import 'package:flutter_boilerplate/features/auth/model/role.dart';
+import 'package:flutter_boilerplate/features/auth/model/user.dart';
+import 'package:flutter_boilerplate/features/auth/repository/auth_repo.dart';
+import 'package:get/get.dart';
+
+class AuthService extends GetxService {
+  final AuthRepo authRepo;
+  AuthService({required this.authRepo}) {
+    loadUser();
+  }
+
+  UserModel? _currentUser;
+  UserModel? get currentUser => _currentUser;
+
+  void loadUser() {
+    final jsonString = authRepo.getUser();
+    if (jsonString != null) {
+      _currentUser = UserModel.fromJson(jsonDecode(jsonString));
+    }
+  }
+
+  Future<void> signUp(UserModel user) async {
+    await authRepo.signUp(jsonEncode(user.toJson()));
+    _currentUser = user;
+    update();
+  }
+
+  Future<void> login(String email, String password) async {
+    final response = await authRepo.login(email, password);
+    if (response.isOk) {
+      loadUser();
+    }
+    update();
+  }
+
+  Future<void> logout() async {
+    await authRepo.logout();
+  }
+
+  bool get isLoggedIn => authRepo.isLoggedIn();
+
+  bool get isOwner => currentUser?.role == UserRole.owner;
+
+  Future<void> saveUser(UserModel user) async {
+    _currentUser = user;
+    await authRepo.saveUser(jsonEncode(user.toJson()));
+    update();
+  }
+}
