+++ lib/features/auth/signup_screen.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_boilerplate/base/custom_button.dart';
 import 'package:flutter_boilerplate/base/custom_text_field.dart';
 import 'package:flutter_boilerplate/features/auth/controller/auth_controller.dart';
-import 'package:flutter_boilerplate/data/model/user/role.dart';
-import 'package:flutter_boilerplate/data/model/user/user.dart';
+import 'package:flutter_boilerplate/features/auth/model/role.dart';
+import 'package:flutter_boilerplate/features/auth/model/user.dart';
 import 'package:flutter_boilerplate/features/dashboard/assistant_dashboard.dart';
 import 'package:flutter_boilerplate/features/dashboard/owner_dashboard.dart';
 import 'package:get/get.dart';
