+++ lib/features/finance/model/wallet_transaction.dart
