+++ lib/features/finance/repository/advance_repo.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import 'package:flutter_boilerplate/data/model/advance/advance.dart';
+import 'package:flutter_boilerplate/features/finance/model/advance.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class AdvanceRepo {
